January 2020 TriadJS Meetup Sharing. (#6)
Post on work done for a recent local JS meetup.

* January 2020 TriadJS Meetup Post.

First draft of the work done for the first meetup of the year at
TriadJS.

* Change Internal Name of New Post.

January seems repetitive when looking at it from netlify's end. Instead
let's focus on the Speaker Series part and not so much when it occurred.

* Rename File.

This will rename/remap the URL. This now aligns better with the title
change

* Rename Title.

Focus on the title of the presentation. Again, this kinda makes more
sense when reading it off from the list of posts. Giving the potential
reader more information prior to clicking into the post.