Post: 2019/2020 Bag Awards (#7)
* V1 Writeup About Tech Pouches.

After finding the awesome sauce that is the Waterfield design's tech
pouch, let's praise it from here to the top of that blrg.

* Second Draft.

Finish off the final and second draft of the post highlighting my
favorite tech pouch of 2019 and highly likely 2019. Mostly includes
fixes to grammar and other spaces where I really went off the rails at
the end of the post.