Three R's Followup Post (#9)
* Install Remark Frontmatter Plugin.

Markdown for Eleventy uses frontmatter. This will get removed when
utilizing remark's CLI on these files. Install this plugin to allow and
configure the usage of frontmatter.

* Add A Tasks File.

Move Remark Linting to the Tasks file. This will make linting/fixing
markdown errors easier, more manageable.

* 1st Draft

Build out a first draft of the three r's post. This so far embodies
really the core of what needs to be said about these three communities.
Not entirely sure if more expounding on the subject is needed... or
what?

* Attempt To Configure the List Indent.

For some reason, the list indent is a bit off in the lint recommended
preset for Remark. Trying to move this to a single space versus a tab
seems to be not working, although this is the configuration that
'should' be correct.

* Last Draft of Three R's