2020 Programming Language of the Year Post (#12)
* First Draft.

Build out a simple and definitely rough first draft.
Little to no linking and certainly few well structured sentences.
Worthwhile start though to the first year in a long time
with no new programming learned or re-learned.

* Finish Initial Links.

Blow the linking game (as always) out of the water.
This introduces a complete (for now) set of links for my
'2020 programming language of the year' post.

* Finish Second Version Of Post.

The second and final version draws this post to a publishable close.
Most of the changes here revolve around passive voice and spelling.
Including a small footnote to see how eleventy is currently displaying
them.

* Wire Up Footnotes When Building Site.

Markdown it, out of the box does not actually render footnotes
correctly. This requires a different package. Install said package and
wire it into the eleventy config. This now allows footnotes to render as
expected.