Start Here Section (#14)
* Begin Building A Start Here Section.

This will live at the very top of the homepage and will service as a
welcome point for folks who might have no idea what's going on. And my
title doesn't give them enough oomph to read more.

* Flush Out More Of The Top Six Section.

* Set First Round Of Top6

Close in on the first top 6. This fills out the first
section/introductory section of the blog.

* Write Up Blog Post Introducing Top 6