Roam Consolidation (#18)
* Roam/Gardening First Draft.

Noodling on an internet garden/roam post about how the tool did a
consolidation number on my writing and daily thinking. Pretty
interesting, I feel but, going to need to give it a couple more once
overs to get an overall hype feeling.

* Final Pass on Roam Post.

Finish up this post that is more of an introduction to Roam in this
space than anything.