Change Of Mindset over 2020 (#23)
Riffing on a few recent blog posts around this topic. We hold a lot of truths as truths as programmers without a lot of thought otherwise. Take a retrospective look of the last couple of years when it comes to how I changed my tune (and perhaps when I didn't)