JSON Feed (#25)
Plugin for managing a JSON like feed (much like xml). Includes a bump
to the version 2 of Package.lock and some risky packages.
There really isn't a need for a plugin here with the JSON feed. So few
readers manage to use it. Instead provide a slimmed down version for
now.