Eleventy 1 Beta (#31)
This bumps the blog to Eleventy 1.0 beta. There have been quite a few lately so this should be a fairly easy fix/bump. Get's the entire project ready for when this inevitably gets the green light.

Also includes some post linting cleanup as well as some changes to the way that the overall blog is built.