Migrate To Just CSS and Lightning. (#60)
* feat(css): Migrate To Just CSS.

Create a few copies of the scss files and move them
over to just css. Eventually with the hope to use
PostCSS to concatenate them into one single file
for the Blog to host.

* chore: Remove Direnv Directory.

Anything that's being built by Nix through direnv,
ignore.

* feat: PostCSS Install.

Grab PostCSS and a few plugins to be able to produce
a solid CSS output for the blog.

* chore(deps): TS Items.

Mostly for editor configs. If an editor expects
access to TS, make sure it's installed via NPM.

* chore(env): Set Node ENV.

With some of the new tooling, NODE_ENV is required
to make sure that the css is readable locally but
smushed when deploying.

* feat(deps) Get Post CSS Working With Tasks.

Most of 11ty is all about those NPM tasks. Wire
up and add the correct PostCSS packages to get the
build correct.

* feat(css) Post CSS Config File.

Add PostCSS config to be able to properly build css
when in development and production.

* feat(css) Wire Up New Index CSS Page.

No need to split these at the moment. There really
isn't that much CSS being passed around. Instead,
opt for an import strategy and add this index to all
of the included liquid templates.

* feat(css) Swap PostCSS For Lightning.

Opt for one of these all-in-one tools that's built
in Rust... By hipsters down by the bay.

* fix: Omit Favicon Errors.

Sort of 'hack the favicon'. No need for these darn
things with a blog. Calm down the browser from
throwing errors in the console. Keeping things
nice and tidy.

* feat: Remove SCSS.

And all of that SCSS is now gone. Opting for more
modern CSS only items + items that Lightning gives
us.

* chore(deps): NPM Audit Fix.

Bump NPM modules to get out of any issues there
maybe with the site and NPM.