Bump To Styles (#68)
* feat: Ignore AI Files.

Utilize Repomix to make working with the repo much
smaller and easier to deal with. There isn't much
going on here but, this makes things smaller/more
compact.

* feat: Tweaks To The Main Body.

Overall the width of the blog has been really weird.
This aligns it with a cleaner, compressed reading
experience.

* fix: Blockquote Styling.

Have always struggled with the whole design of the
blockquotes. Giving a bigger, more promenant one a
whirl.

* fix: Style Links/Buttons.

Some changes to make the long list of links on the
index a bit more nice? Somewhat. There's likely
a better way here but, this is an incremental or
'non jarring' change.