feat: Minify CSS In Production.
Use CleanCSS to minify the code in production.
Utilize the transform functions available to the
pluginBundle.

The function makes sure that css is the type and
that the correct env variable is set. If everything
is good, minify and clean the code. This will produce
a much smaller/svelter production set of inline code.