Fingerprinting CSS Assets. (#69)
* feat: Lightning CSS Plugin For 11ty

Utilize a plugin over concurrently/double build
steps when serving the application. This also drastically
speeds up the builds considerably.

This fix currently only works in dev.

* feat: Add A Script For Manipulating CSS.

This script will take the expected CSS and create
a hashed and compressed file for serving over Cloudflare
and a normal CSS file in development.

* feat: Utilize New CSS File in Templates.

Add this new registered function into the eleventy
config and utilize the output in special hooks in
the templates.

* refactor: Fix/Remove Use of Lightning CSS Imports.

The imports need a bit of fixing/cutting and the
package.json still references some old code that's
no longer needed.

* feat: Lightning CSS In Brotli Helper.

Instead of using CleanCSS to transform the CSS,
utilize Lightning. This dependency was always hanging
around and ultimately it's the best way forward with
CSS at the moment.

Utilize it's ability to minimize the CSS when being
deployed to prod.

* fix: Builds Without Explicit Lightning Scripts.

Clean up the manual Lightning builds for both the
dev and production deploys.

* feat: ES6 Module This Thing Everywhere.

Move this all over to a module approach using es6
throughout the eleventy config.

* feat: Output Cache Key In Logs.

This console.log should run when the application
initially builds the cache key. This should aid
in debugging why this isn't being created correctly
on Cloudflare.

* chore: Ignore Manifest Files.

Manifest files for CSS busting. Ignore these locally
since they will be run in development. Production
will inline the css for speed and POWER.

* feat: Opt For the Bundle Plugin.

This is now built into 11ty as of version 3. This
bundling allows us to build custom bundles of CSS/JS
for the application.

In this case, we bundle all of the imports together
during DEV into a single file and inline in production.

* feat: Minify CSS In Production.

Use CleanCSS to minify the code in production.
Utilize the transform functions available to the
pluginBundle.

The function makes sure that css is the type and
that the correct env variable is set. If everything
is good, minify and clean the code. This will produce
a much smaller/svelter production set of inline code.