refactor(post): Fixing Up The Solid V3/V4 Version.
This short little post has been the first full
one to utilize Vale and the LSP work around writing
that exists in this repo now.

Due to this, the time it takes to get something through
the editorial process seems to be much longer than normal.
Finishing this one now.