Confusing Diff Views Break My Brain (Post) (#73)
* feat: Brand New V1 Post Around Diffviews.

Either my eyesight is getting bad or really,
the visuals of GitLab are a bit poo.

* fix: V2 of Editing Process.

Fix a lot of linting issues from Vale but, also some
structural issues from other tools like Marked.

* refactor(post): Fixing Up The Solid V3/V4 Version.

This short little post has been the first full
one to utilize Vale and the LSP work around writing
that exists in this repo now.

Due to this, the time it takes to get something through
the editorial process seems to be much longer than normal.
Finishing this one now.